RavenDB-18204 : fixes after rebase : use DatabaseContext.Operations instead of Server.Operations, change StartBackupOperationResult.OperationId from int to long

+++ src/Raven.Client/Documents/Operations/Backups/StartBackupOperation.cs
@@ -64,6 +64,6 @@ namespace Raven.Client.Documents.Operations.Backups
     {
         public string ResponsibleNode { get; set; }
 
-        public int OperationId { get; set; }
+        public long OperationId { get; set; }
     }
 }
